move config file and models folder
