+++ src/dao/KeywordDao.js
@@ -5,7 +5,7 @@ const {
     sequelize,
     JoinPerfumeKeyword,
     JoinReviewKeyword,
-} = require('../../models');
+} = require('../models');
 const { Op } = Sequelize;
 
 /**
