+++ src/dao/PerfumeDao.js
@@ -12,10 +12,10 @@ const {
     SearchHistory,
     sequelize,
     Sequelize,
-} = require('../../models');
+} = require('../models');
 const { Op } = Sequelize;
 
-const { ranking } = require('../../mongoose_models');
+const { ranking } = require('../mongoose_models');
 
 const PERFUME_THUMB_COLUMNS = [
     'perfumeIdx',
