+++ src/dao/ReviewDao.js
@@ -10,7 +10,7 @@ const {
     JoinReviewKeyword,
     JoinPerfumeKeyword,
     Keyword,
-} = require('../../models');
+} = require('../models');
 const { Op } = Sequelize;
 
 /**
