+++ src/models/searchHistory.js
