+++ src/models/series.js
