+++ test/dao/KeywordDao.spec.js
@@ -3,7 +3,7 @@ dotenv.config();
 
 const keywordDao = require('../../src/dao/KeywordDao');
 const { expect } = require('chai');
-const { Sequelize } = require('../../models');
+const { Sequelize } = require('../../src/models');
 const { Op } = Sequelize;
 
 describe('# KeywordDao Test', () => {
