+++ test/dao/ReviewDao.spec.js
@@ -5,7 +5,7 @@ const chai = require('chai');
 const { expect } = chai;
 const reviewDao = require('../../src/dao/ReviewDao.js');
 const keywordDao = require('../../src/dao/KeywordDao');
-const { Review, JoinReviewKeyword } = require('../../models');
+const { Review, JoinReviewKeyword } = require('../../src/models');
 const { read } = require('../../src/dao/NoteDao.js');
 const { NotMatchedError } = require('../../src/utils/errors/errors.js');
